+++ src/main/java/com/dfsek/paralithic/eval/ParserUtil.java
@@ -1,16 +1,16 @@
 package com.dfsek.paralithic.eval;
 
-import com.dfsek.paralithic.operations.Node;
-import com.dfsek.paralithic.operations.binary.*;
-import com.dfsek.paralithic.operations.binary.booleans.AndNode;
-import com.dfsek.paralithic.operations.binary.booleans.ComparisonNode;
-import com.dfsek.paralithic.operations.binary.booleans.OrNode;
-import com.dfsek.paralithic.operations.binary.number.AdditionNode;
-import com.dfsek.paralithic.operations.binary.number.DivisionNode;
-import com.dfsek.paralithic.operations.binary.number.ModuloNode;
-import com.dfsek.paralithic.operations.binary.number.MultiplicationNode;
-import com.dfsek.paralithic.operations.binary.number.SubtractionNode;
-import com.dfsek.paralithic.operations.binary.special.PowerNode;
+import com.dfsek.paralithic.node.Node;
+import com.dfsek.paralithic.node.binary.*;
+import com.dfsek.paralithic.node.binary.booleans.AndNode;
+import com.dfsek.paralithic.node.binary.booleans.ComparisonNode;
+import com.dfsek.paralithic.node.binary.booleans.OrNode;
+import com.dfsek.paralithic.node.binary.number.AdditionNode;
+import com.dfsek.paralithic.node.binary.number.DivisionNode;
+import com.dfsek.paralithic.node.binary.number.ModuloNode;
+import com.dfsek.paralithic.node.binary.number.MultiplicationNode;
+import com.dfsek.paralithic.node.binary.number.SubtractionNode;
+import com.dfsek.paralithic.node.binary.special.PowerNode;
 
 public class ParserUtil {
     public static BinaryNode createBinaryOperation(BinaryNode.Op op, Node left, Node right) {
