+++ src/main/java/com/dfsek/paralithic/eval/parser/Parser.java
@@ -21,14 +21,14 @@ import com.dfsek.paralithic.functions.natives.NativeFunction;
 import com.dfsek.paralithic.functions.natives.NativeMath;
 import com.dfsek.paralithic.functions.operation.OperationFunction;
 import com.dfsek.paralithic.functions.operation.TernaryIfFunction;
-import com.dfsek.paralithic.operations.Constant;
-import com.dfsek.paralithic.operations.special.InvocationVariableNode;
-import com.dfsek.paralithic.operations.Node;
-import com.dfsek.paralithic.operations.binary.BinaryNode;
-import com.dfsek.paralithic.operations.special.function.FunctionNode;
-import com.dfsek.paralithic.operations.special.function.NativeFunctionNode;
-import com.dfsek.paralithic.operations.unary.AbsoluteValueNode;
-import com.dfsek.paralithic.operations.unary.NegationNode;
+import com.dfsek.paralithic.node.Constant;
+import com.dfsek.paralithic.node.special.InvocationVariableNode;
+import com.dfsek.paralithic.node.Node;
+import com.dfsek.paralithic.node.binary.BinaryNode;
+import com.dfsek.paralithic.node.special.function.FunctionNode;
+import com.dfsek.paralithic.node.special.function.NativeFunctionNode;
+import com.dfsek.paralithic.node.unary.AbsoluteValueNode;
+import com.dfsek.paralithic.node.unary.NegationNode;
 
 import java.io.Reader;
 import java.io.StringReader;
