+++ src/main/java/com/dfsek/paralithic/functions/operation/TernaryIfFunction.java
@@ -1,7 +1,7 @@
 package com.dfsek.paralithic.functions.operation;
 
-import com.dfsek.paralithic.operations.Node;
-import com.dfsek.paralithic.operations.special.TernaryIfNode;
+import com.dfsek.paralithic.node.Node;
+import com.dfsek.paralithic.node.special.TernaryIfNode;
 
 import java.util.List;
