+++ src/main/java/com/dfsek/paralithic/node/Node.java
@@ -1,4 +1,4 @@
-package com.dfsek.paralithic.operations;
+package com.dfsek.paralithic.node;
 
 import org.jetbrains.annotations.NotNull;
 import org.objectweb.asm.MethodVisitor;
