+++ src/main/java/com/dfsek/paralithic/node/OperationUtils.java
@@ -1,4 +1,4 @@
-package com.dfsek.paralithic.operations;
+package com.dfsek.paralithic.node;
 
 import org.objectweb.asm.MethodVisitor;
 import static org.objectweb.asm.Opcodes.*;
