+++ src/main/java/com/dfsek/paralithic/node/binary/booleans/ComparisonNode.java
@@ -1,8 +1,8 @@
-package com.dfsek.paralithic.operations.binary.booleans;
+package com.dfsek.paralithic.node.binary.booleans;
 
-import com.dfsek.paralithic.operations.Node;
-import com.dfsek.paralithic.operations.binary.BinaryNode;
-import com.dfsek.paralithic.operations.Constant;
+import com.dfsek.paralithic.node.Node;
+import com.dfsek.paralithic.node.binary.BinaryNode;
+import com.dfsek.paralithic.node.Constant;
 import org.jetbrains.annotations.NotNull;
 import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
