+++ src/main/java/com/dfsek/paralithic/node/special/TernaryIfNode.java
@@ -1,9 +1,9 @@
-package com.dfsek.paralithic.operations.special;
+package com.dfsek.paralithic.node.special;
 
-import com.dfsek.paralithic.operations.Node;
-import com.dfsek.paralithic.operations.OperationUtils;
-import com.dfsek.paralithic.operations.Simplifiable;
-import com.dfsek.paralithic.operations.Constant;
+import com.dfsek.paralithic.node.Node;
+import com.dfsek.paralithic.node.OperationUtils;
+import com.dfsek.paralithic.node.Simplifiable;
+import com.dfsek.paralithic.node.Constant;
 import org.jetbrains.annotations.NotNull;
 import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
