+++ src/main/java/com/dfsek/paralithic/node/unary/AbsoluteValueNode.java
@@ -1,7 +1,7 @@
-package com.dfsek.paralithic.operations.unary;
+package com.dfsek.paralithic.node.unary;
 
-import com.dfsek.paralithic.operations.Node;
-import com.dfsek.paralithic.operations.Constant;
+import com.dfsek.paralithic.node.Node;
+import com.dfsek.paralithic.node.Constant;
 import org.jetbrains.annotations.NotNull;
 import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
